+++ CHANGELOG.md
@@ -15,7 +15,7 @@ any parts of the framework not mentioned in the documentation should generally b
 * Moved resolving of `included_serialzers` and `related_serializers` classes to serializer's meta class.
 * `AutoPrefetchMixin` updated to be more clever about how relationships are prefetched, with recursion all the way down.
 * Expensive reverse relations are now automatically excluded from queries that don't explicitly name them in sparsefieldsets. Set `INCLUDE_EXPENSVE_FIELDS` to revert to old behaviour.
-* Removed `PreloadIncludesMixin`, as the logic did not work when nesting includes, and the laborious effort needed in its manual config was unnecessary.
+* Removed `PreloadIncludesMixin`, as the logic did not work when nesting includes, and the laborious effort needed in its manual config was unnecessary. This removes support for `prefetch_for_includes` and `select_for_includes`
 
 ### Deprecated
